+++ LS_RXHelper/lpred.cpp
@@ -1,4 +1,5 @@
 #include "stdafx.h"
+#include "AntiAFK.h"
 
 PSDK_CONTEXT SDK_CONTEXT_GLOBAL;
 
@@ -17,34 +18,10 @@ SDKCOLOR _g_ColorPurple = { 128, 0, 128, 255 };
 
 	 bool showwindow;
 	 bool bPred;
+	 bool bAntiAFK;
 
 }m_Options;
 
- EEvent<void, int, int>testevent_args;
- EEvent<void>testevent_noargs;
- void OnTestEventArgs(int arg1, int arg2)
- {
-	 printf("Args : %d,%d\n", arg1, arg2);
- }
- void OnTestEvent_NO_Args()
- {
-	 printf("NO args\n");
- }
-
- // conformance option should be disabled
- int eventexample(void)
- {
-	 bindEEvent(testevent_args, OnTestEventArgs, 1, 2);
-	 int eventid2 = bindEEvent(testevent_noargs, OnTestEvent_NO_Args);
-
-
-	 notifyEEvent(testevent_args, 1, 2); // with infinite args
-	 notifyEEvent(testevent_noargs); // no args
-
-	 if (GetAsyncKeyState(VK_SPACE))
-		 unbindEEvent(testevent_noargs, eventid2);
-	 return 0;
- }
 
 
 BOOL 
@@ -101,6 +78,7 @@ DllMain(
 	// to render.
 	//
 	SdkRegisterGameScene(DrawGameScene, NULL);
+	SdkRegisterGameScene(AntiAFKTick, NULL);
 
 	//
 	// The base 'game frame' sits below our extra visuals (the 'game
@@ -111,6 +89,14 @@ DllMain(
 }
 
 
+void __cdecl AntiAFKTick(void* UserData) {
+	UNREFERENCED_PARAMETER(UserData);
+
+	if (m_Options.bAntiAFK) {
+		AntiAFK::Execute();
+	}
+}
+
 void __cdecl DrawOverlayScene(_In_ void* UserData)
 {
 	UNREFERENCED_PARAMETER(UserData);
@@ -120,33 +106,12 @@ void __cdecl DrawOverlayScene(_In_ void* UserData)
 	if (ObjectExpanded)
 	{
 		SdkUiCheckbox("Test", &m_Options.bPred, NULL);
+		SdkUiCheckbox("AntiAFK", &m_Options.bAntiAFK, NULL);
 		SdkUiEndTree();
 	}
 
 	
 }
-
-
-double CalcAttackTime()
-{
-	typedef float(__cdecl* fnGetAttackDelay)(void* pObj, int index);
-	fnGetAttackDelay stGet = reinterpret_cast<fnGetAttackDelay>((DWORD)(GetModuleHandleA(0)) + 0x64C5C0); // 1.  57 8B 7C 24 08 8B 87 ? ? ? ? 8D 8F ? ? ? ?
-	float delay = stGet(EntityManager::GetLocalPlayer().GetObjectPTR(), 64); 
-	float ChampionAttackSpeed = 1.0f / delay; // Like 2.50 
-	float Div = EntityManager::GetLocalPlayer().GetAttackSpeed(); // like 4.12
-	float BaseMult = ChampionAttackSpeed / Div; //Like 0.600 
-	return BaseMult;
-}
-double CalcWindup()
-{
-	typedef float(__cdecl* fnGetAttackDelay)(void* pObj, int index);
-	fnGetAttackDelay stGet = reinterpret_cast<fnGetAttackDelay>((DWORD)(GetModuleHandleA(0)) + 0x64C440); // 83 ec ? 53 8b 5c ? ? 8b cb 56 ---- 2.  57 8B 7C 24 08 8B 87 ? ? ? ? 8D 8F ? ? ? ?
-	auto delay = stGet(EntityManager::GetLocalPlayer().GetObjectPTR(), 1);
-	float pDelay = delay * 2 ;
-	SdkUiConsoleWrite("pDelay : %f delay : %f from rift attackspeed : %f", pDelay, delay, EntityManager::GetLocalPlayer().GetAttackSpeed());
-	return pDelay;
-}
-
 
 void __cdecl DrawGameScene( _In_ void* UserData)
 {
@@ -156,21 +121,13 @@ void __cdecl DrawGameScene( _In_ void* UserData)
 	// Update entitymanager
 	//
 
-    // 
-	// event examples
-	// 
-
 	EntityManager::Update();
+
+
 	if (!m_Options.bPred)
 		return;
 	
-	if (GetAsyncKeyState(VK_SPACE))
-	{
-		auto windup = CalcWindup();
-		auto attdelay = CalcAttackTime();
-		//SdkUiConsoleWrite("windup : %f attack delay : %f", windup,attdelay);
-	}
-	
+
 
 	/*
 	BUFF EXAMPLE
@@ -197,9 +154,7 @@ void __cdecl DrawGameScene( _In_ void* UserData)
 			auto Pos = hero.GetPosition();
 			Pos.z += (i * 15);
 
-			SdkDrawText(&Pos, NULL, buff.Name, "Arial", &_g_ColorWhite, 18, 5, 0, false);
-
-			
+			SdkDrawText(&Pos, NULL, buff.Name, "Arial", &_g_ColorWhite, 18, 5, 0, false);	
 		}
 	}
